requirements.txt
fastapi
uvicorn
python-dotenv

# Crawling
playwright
chromium

# RAG & Embeddings
sentence-transformers
faiss-cpu

# LangChain + Groq
langchain
langchain-community
langchain-groq
langchain-text-splitters
gunicorn
